make abstract class AckWithTimeOut that implements Ack and update tests
